[refactor]: dedup studio blob downloads, drop dead UploadZone mode
Hoist the blob-download anchor dance into a shared downloadBlob helper
(was written three times with drift: DownloadCaptions skipped the
appendChild Firefox needs) and remove UploadZone's textInput mode,
which no caller ever used.
